pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["nguyen <nguyenhainguyen2606@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = "^2.2.1"
pypdf = "^4.1.0"
selenium = "^4.18.1"
beautifulsoup4 = "^4.12.3"
translators = "^5.9.0"
faster-whisper = "^1.0.1"
numpy = "^1.26.4"
tiktoken = "^0.6.0"
pymongo = "<=4.6.1"
ndjson = "^0.3.1"
google-cloud-bigquery = "<3.13.0"
google-cloud-core= "<2.3.3"
apache-airflow = "^2.8.3"
google-cloud = "^0.34.0"
google = "^3.0.0"
include = "^0.2.2"
gcloud = "^0.18.3"
gcloud-aio-auth = "^5.2.2"
ruff = "^0.3.5"
boto3 = "^1.34.76"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
